backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bitarray==3.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
cytoolz==1.0.1
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.2.0
eth-abi==5.2.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
frozenlist==1.5.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.5
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
parsimonious==0.10.0
passlib==1.7.4
pip==24.0
propcache==0.2.1
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyjwt==2.10.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pyunormalize==16.0.0
pywin32==308
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rlp==4.1.0
rsa==4.9
setuptools==65.5.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
toolz==1.0.0
typer==0.15.1
types-requests==2.32.0.20241016
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
web3==7.8.0
websockets==13.1
yarl==1.18.3
